extract out util functions to separate file
